requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.16.0
click==8.1.7
ecdsa==0.18.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
mypy==1.8.0
mypy-extensions==1.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
